// src/app/models/initialRegis.ts
export interface InitialRegister {
    //id: string;
  firstName: string | null;
  lastName: string | null;
  email: string | null;
  password: string | null;
  dateOfBirth: string | null;
  countryOfResidence: string | null;
  postalCode: string | null;
  contactNumber: string | null;
  //accountType: ProfileType;
  //shipments: [number | null] | null;
  }